requirements.txt
appnope==0.1.2
async-generator==1.10
attrs==20.3.0
backcall==0.2.0
bleach==3.2.1
certifi==2020.6.20
chardet==3.0.4
decorator==4.4.2
defusedxml==0.6.0
entrypoints==0.3
idna==2.10
influxdb==5.3.0
iniconfig==1.1.1
ipykernel==5.4.2
ipython==7.19.0
ipython-genutils==0.2.0
jedi==0.17.2
Jinja2==2.11.2
jsonschema==3.2.0
jupyter-client==6.1.7
jupyter-core==4.7.0
jupyterlab-pygments==0.1.2
Markdown==3.3.3
MarkupSafe==1.1.1
minimalmodbus==1.0.2
mistune==0.8.4
msgpack==0.6.1
nbclient==0.5.1
nbconvert==6.0.7
nbformat==5.0.8
nest-asyncio==1.4.3
numpy==1.19.4
packaging==20.8
pandas==1.1.5
pandocfilters==1.4.3
parso==0.7.1
pexpect==4.8.0
pickleshare==0.7.5
pika==1.1.0
pluggy==0.13.1
prompt-toolkit==3.0.8
ptyprocess==0.6.0
Pweave==0.30.3
py==1.10.0
Pygments==2.7.3
pymongo==3.11.0
pyparsing==2.4.7
pyrsistent==0.17.3
pyserial==3.5
pytest==6.1.1
python-dateutil==2.8.1
pytz==2020.1
PyYAML==5.3.1
pyzmq==20.0.0
redis==3.5.3
requests==2.24.0
six==1.15.0
testpath==0.4.4
toml==0.10.2
tornado==6.1
traitlets==5.0.5
urllib3==1.25.11
wcwidth==0.2.5
webencodings==0.5.1
certifi==2020.6.20
chardet==3.0.4
idna==2.10
influxdb==5.3.0
msgpack==0.6.1
pika==1.1.0
pymongo==3.11.0
python-dateutil==2.8.1
pytz==2020.1
PyYAML==5.3.1
redis==3.5.3
requests==2.24.0
six==1.15.0
urllib3==1.25.11
pytest==6.1.1

